requirements.txt
click>=7.1.2
numpy>=1.24.4
opencv_python>=4.1.2.30
torch>=1.9.0
imgaug>=0.4.0
termcolor>=1.1.0
torchvision>=0.14.1
matplotlib>=3.7.2
path.py>=12.5.0
PyYAML>=6.0.1
